refactor(autoTagger): move autoTagger to src directory
- Relocate autoTagger.ts from root to src directory
- Update import paths to reflect new location
- Maintain existing functionality without changes
